+++ include/ccf/byte_vector.h
@@ -20,9 +20,9 @@ namespace std
   {
     size_t operator()(const llvm_vecsmall::SmallVector<T, N>& v) const
     {
-      static constexpr siphash::SipKey k{
+      static constexpr ccf::siphash::SipKey k{
         0x7720796f726c694b, 0x2165726568207361};
-      return siphash::siphash<2, 4>(v.data(), v.size(), k);
+      return ccf::siphash::siphash<2, 4>(v.data(), v.size(), k);
     }
   };
 }
