+++ include/ccf/ds/hash.h
@@ -42,9 +42,9 @@ namespace std
     {
       // For cryptographically secure hashing, use SipHash directly with a
       // secret key. For std::hash, we use this fixed key
-      static constexpr siphash::SipKey k{
+      static constexpr ccf::siphash::SipKey k{
         0x7720796f726c694b, 0x2165726568207361};
-      return siphash::siphash<2, 4>(v, k);
+      return ccf::siphash::siphash<2, 4>(v, k);
     }
   };
