+++ src/enclave/thread_local.cpp
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the Apache 2.0 License.
-#include "ccf/ds/thread_ids.h"
+#include "ccf/threading/thread_ids.h"
 
 namespace ccf::threading
 {
